Fix #9 Include wiki
A second GitLocal instance is used to control the wiki repository.  Unfortunately the GitHub API does not provide a way to create the first wiki page, which means that manual intervention is required until that first page is created.

ChangeLogConfiguration simplifies the selection of a target output directory with the  ```OutputTarget``` enum - which includes options such as PROJECT_ROOT, WIKI_ROOT and others