Fix #46 GitPlus parameterless constructor
Fixed a problem in .gitignore which makes 0.9.14 and 0.9.15 invalid - 'build' in .gitignore was excluding the entire 'build' package, as it was missing the leading '/' - This causes a lot of files to be added, as they had previously been deleted by refactoring but not replaced.

GitPlus construction changed, now has only parameterless constructor, with GitLocalProvider passed in as part of configuration (DefaultGitLocalProvider as default).
GitPlusConfiguration is created when getConfiguration() is called.
Validate method in GitPlusConfiguration has been removed, and its functionality spread across the various getters so checks are carried out only when that part of the configuration is used.