Fix #147 Dependencies updated, all now at latest.  Tests amended to reflect a change in clone behaviour
Tests which previously passed (even when run on earlier versions) were failing becuase of a change in behaviour.  Previously, a clone would result in having the default branch checked out, but it seems now that it is the 'master' branch which is checked out, even if, say, 'develop' is the default.  The puzzle is how or when this changed